refactor(shape): satisfy the code-shape ratchet on the parity branch
CI's quality:shape gate flagged three growths: importTasks reached 41
lines with the cap guard, edit-drawer-fields.tsx crossed the file cap at
460 when the reminder control landed, and useEditDraftState grew to 61
lines. Pure extractions, no behavior change:

- the import-cap guard moves to assertWithinImportCap()
- the reminder control moves to edit-drawer-reminder-field.tsx
- the due-date state trio moves to a useDueDateDraft() helper hook
